# spec/spec/cortex/openapi/cortex.openapi.yaml
openapi: 3.0.0
info:
  title: LOXA Cortex API
  version: v1
  description: Incident detection and reconstruction API

servers:
  - url: https://api.loxa.dev
    description: Production
  - url: http://localhost:8080
    description: Local development

paths:
  /healthz:
    get:
      summary: Health check
      operationId: healthz
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string

  /v1/events:
    post:
      summary: Ingest a single event
      operationId: ingestEvent
      tags:
        - Ingest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CortexEvent'
      responses:
        '202':
          description: Event accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestEventResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

  /v1/events/batch:
    post:
      summary: Ingest batch of events
      operationId: ingestBatch
      tags:
        - Ingest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CortexEvent'
      responses:
        '202':
          description: Batch processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestBatchResponse'

  /v1/incidents/{incident_id}/reconstruct:
    post:
      summary: Reconstruct incident timeline
      operationId: reconstruct
      tags:
        - Reconstruction
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                mode:
                  type: string
                  enum: [fast, deep]
                  default: fast
      responses:
        '200':
          description: Reconstruction complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconstructResponse'

  /v1/graph/service/{service}:
    get:
      summary: Get service dependency graph
      operationId: getServiceGraph
      tags:
        - Graph
      parameters:
        - name: service
          in: path
          required: true
          schema:
            type: string
        - name: depth
          in: query
          schema:
            type: integer
            default: 2
      responses:
        '200':
          description: Service graph
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphResponse'

  /v1/graph/incident/{incident_id}:
    get:
      summary: Get incident graph
      operationId: getIncidentGraph
      tags:
        - Graph
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Incident graph
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphResponse'

  /v1/feedback/remediation:
    post:
      summary: Record remediation feedback
      operationId: recordFeedback
      tags:
        - Feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
      responses:
        '201':
          description: Feedback recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string

components:
  schemas:
    CortexEvent:
      type: object
      required:
        - event_id
        - timestamp
        - service
        - event
        - kind
        - provenance
      properties:
        event_id:
          type: string
        timestamp:
          type: string
          format: date-time
        service:
          type: string
        event:
          type: string
        kind:
          type: string
          enum:
            - event
            - http
            - job
            - queue
            - cli
            - cron
            - log
            - checkpoint
            - deploy
            - metric
            - trace
            - topology
            - incident_signal
            - remediation
            - loxa_event
            - otel_log
            - otel_span
            - collector_event
        level:
          type: string
          enum: [debug, info, warning, error, critical]
        outcome:
          type: string
          enum: [success, failure, partial, unknown]
        trace_id:
          type: string
        incident_id:
          type: string
        provenance:
          type: string
          enum: [loxa, collector, otlp, jsonl, manual, replay]
        raw:
          type: object

    IngestEventResponse:
      type: object
      properties:
        status:
          type: string
        event_id:
          type: string
        incident_id:
          type: string

    IngestBatchResponse:
      type: object
      properties:
        status:
          type: string
        accepted:
          type: integer
        rejected:
          type: integer
        invalid:
          type: integer
        errors:
          type: array
          items:
            $ref: '#/components/schemas/IngestError'

    IngestError:
      type: object
      properties:
        event_id:
          type: string
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean

    ValidationError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string

    GraphNode:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        label:
          type: string
        service:
          type: string
        attributes:
          type: object

    GraphEdge:
      type: object
      properties:
        id:
          type: string
        from_node_id:
          type: string
        to_node_id:
          type: string
        type:
          type: string
        weight:
          type: number

    GraphResponse:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/GraphNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/GraphEdge'

    ReconstructResponse:
      type: object
      properties:
        incident_id:
          type: string
        status:
          type: string
        severity:
          type: string
        primary_service:
          type: string
        affected_services:
          type: array
          items:
            type: string
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/GraphNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/GraphEdge'
        evidence:
          type: array
          items:
            type: object
            properties:
              event_id:
                type: string
              reason:
                type: string
              confidence:
                type: number
        summary:
          type: string
        confidence:
          type: number

    FeedbackRequest:
      type: object
      required:
        - incident_id
        - remediation_id
        - success
      properties:
        incident_id:
          type: string
        remediation_id:
          type: string
        success:
          type: boolean
        outcome:
          type: string
        time_to_resolve_seconds:
          type: integer
        notes:
          type: string

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer

security:
  - BearerAuth: []